ServiciosPublicos25

+++ atogobmx-front/src/views/ServiciosPublicos/streetLighting/expedientlighting/ExpedientLightingList.vue
@@ -11,15 +11,17 @@
       <b-button
         variant="primary"
         style="
+          background-color: rgb(94,80,238);
           height: 50px;
           width: auto;
           font-size: 18px;
           margin-right: 15px;
           margin-left: 20px;
         "
-        v-b-modal.modal-expedientlighting
+        @click="showModal = !showModal"
         type="submit"
       >
+        <!-- v-b-modal.modal-expedientlighting -->
         <i class="bi bi-person-plus-fill"></i>
         Agregar Expediente
       </b-button>
@@ -64,32 +66,50 @@
       title="Agregar Expediente Alumbrado"
       size="xl"
       hide-footer
+      centered
       button-size="lg"
       lazy
     >
       <Form @submit="addExpedientLighting">
         <b-row cols="2">
+          <!--1-->
           <b-col>
-            <b-form-group class="mt-3" label="Nombre del tipo Tarea">
-              <Field name="TypeTaskField" :rules="validateTypeTask">
-                <b-form-select
-                  v-model="expedientLightingFields.tarea"
-                  autofocus
-                  :state="TypeTaskState"
-                  :options="statusPublicLighting"
-                  value-field="estatusId"
-                  text-field="nombreEstatus"
-                ></b-form-select>
+            <b-form-group class="mt-3" label="Lugar publico">
+              <Field
+                name="PublicPlaceField"
+                :rules="validatePublicPlace"
+                as="text"
+              >
+                <b-form-input
+                  v-model="expedientLightingFields.lugarPublico"
+                  :state="PublicPlaceState"
+                ></b-form-input>
               </Field>
-              <ErrorMessage name="TypeTaskField"
-                ><span>Este campo es requerido </span
-                ><i class="bi bi-exclamation-circle"></i>
-              </ErrorMessage>
+              <ErrorMessage
+                class="text-danger"
+                name="PublicPlaceField"
+              ></ErrorMessage>
             </b-form-group>
           </b-col>
+          <!--2-->
+          <b-col>
+            <b-form-group class="mt-3" label="Localidad">
+              <Field name="LocationField" :rules="validateLocation" as="text">
+                <b-form-input
+                  v-model="expedientLightingFields.localidad"
+                  :state="LocationState"
+                ></b-form-input>
+              </Field>
+              <ErrorMessage
+                class="text-danger"
+                name="LocationField"
+              ></ErrorMessage>
+            </b-form-group>
+          </b-col>
+          <!--3-->
           <b-col>
             <b-form-group class="mt-3" label="Fecha Alta">
-              <Field name="HighDateField" :rules="validateHighDate">
+              <Field name="HighDateField" :rules="validateHighDate" as="">
                 <Datepicker
                   locale="es"
                   name="date"
@@ -98,66 +118,59 @@
                   :state="HighDateState"
                 ></Datepicker>
               </Field>
-              <ErrorMessage name="HighDateField"
-                ><span>Este campo es requerido </span
-                ><i class="bi bi-exclamation-circle"></i>
-              </ErrorMessage>
+              <ErrorMessage name="HighDateField"></ErrorMessage>
             </b-form-group>
           </b-col>
+          <!--4-->
           <b-col>
-            <b-form-group class="mt-3" label="Domicilio">
-              <Field name="DomicileField" :rules="validateDomicile">
-                <b-form-input
-                  v-model="expedientLightingFields.domicilio"
-                  :state="DomicileState"
-                ></b-form-input>
-              </Field>
-              <ErrorMessage name="DomicileField"
-                ><span>Este campo es requerido </span
-                ><i class="bi bi-exclamation-circle"></i>
-              </ErrorMessage>
-            </b-form-group>
-          </b-col>
-          <b-col>
-            <b-form-group class="mt-3" label="Fecha Alta">
-              <Field name="LowDateField" :rules="validateLowDate">
-                <Datepicker
-                  locale="es"
-                  name="date"
-                  text-input
-                  v-model="expedientLightingFields.fechaBaja"
-                  :state="LowDateState"
-                ></Datepicker>
-              </Field>
-              <ErrorMessage name="LowDateField"
-                ><span>Este campo es requerido </span
-                ><i class="bi bi-exclamation-circle"></i>
-              </ErrorMessage>
-            </b-form-group>
-          </b-col>
-          <b-col>
-            <b-form-group class="mt-3" label="Descripcion Domicilio">
+            <b-form-group class="mt-3" label="Departamento">
               <Field
-                name="addresdescriptionField"
-                :rules="validateAddresdescription"
+                name="DepartamentField"
+                :rules="validateDepartament"
+                as="number"
               >
-                <b-form-textarea
-                  v-model="expedientLightingFields.descripcionDomicilio"
-                  :state="addresdescriptionState"
-                  rows="4"
-                ></b-form-textarea>
+                <b-form-select
+                  v-model="expedientLightingFields.departamentoId"
+                  autofocus
+                  :options="departaments"
+                  value-field="departamentoId"
+                  text-field="nombre"
+                  :state="departamentState"
+                  @input="getAreas(EmployeesFields.departamentoId)"
+                >
+                </b-form-select>
               </Field>
-              <ErrorMessage name="addresdescriptionField"
-                ><span>Este campo es requerido </span
-                ><i class="bi bi-exclamation-circle"></i>
-              </ErrorMessage>
+              <ErrorMessage
+                class="text-danger"
+                name="DepartamentField"
+              ></ErrorMessage>
             </b-form-group>
           </b-col>
+          <!--5-->
+          <b-col>
+            <b-form-group class="mt-3" label="Area">
+              <Field name="AreaField" :rules="validateArea" as="number">
+                <b-form-select
+                  v-model="expedientLightingFields.areaId"
+                  autofocus
+                  :options="areas"
+                  value-field="areaId"
+                  text-field="nombre"
+                  :state="areaState"
+                  @input="getWorkStation(EmployeesFields.areaId)"
+                >
+                </b-form-select>
+              </Field>
+              <ErrorMessage class="text-danger" name="AreaField"></ErrorMessage>
+            </b-form-group>
+          </b-col>
+          <!--6-->
           <b-col>
             <b-form-group class="mt-3" label="Descripcion Solucion">
               <Field
                 name="DescriptionSolutionField"
                 :rules="validateDescriptionSolution"
+                as="text"
               >
                 <b-form-textarea
                   v-model="expedientLightingFields.descripcionSolucion"
@@ -165,10 +178,10 @@
                   rows="4"
                 ></b-form-textarea>
               </Field>
-              <ErrorMessage name="DescriptionSolutionField"
-                ><span>Este campo es requerido </span
-                ><i class="bi bi-exclamation-circle"></i>
-              </ErrorMessage>
+              <ErrorMessage
+                class="text-danger"
+                name="DescriptionSolutionField"
+              ></ErrorMessage>
             </b-form-group>
           </b-col>
         </b-row>
@@ -176,8 +189,9 @@
           <b-button
             class="w-auto m-2 text-white"
             variant="primary"
-            v-b-modal.modal-expedientlighting
+            @click="resetexpedientLightingFields"
           >
+            <!-- v-b-modal.modal-expedientlighting -->
             Cancelar
           </b-button>
           <b-button class="w-auto m-2" variant="success" type="submit">
@@ -191,11 +205,13 @@
 
 <script>
 import ExpedientlightingServices from '@/Services/expedientlighting.Services'
-import StatusLightingServices from '@/Services/statuslighting.Services'
+// import StatusLightingServices from '@/Services/statuslighting.Services'
+import AreaServices from '@/Services/area.Services'
+import DepartamentServices from '@/Services/departament.Services'
 import Datepicker from '@vuepic/vue-datepicker'
 import { Form, Field, ErrorMessage } from 'vee-validate'
-import { ref } from 'vue'
-import { useToast } from 'vue-toast-notification'
+import { ref, inject } from 'vue'
+// import { useToast } from 'vue-toast-notification'
 import '@vuepic/vue-datepicker/dist/main.css'
 export default {
   components: {
@@ -206,15 +222,17 @@ export default {
     ErrorMessage
   },
   setup () {
-    const {
-      getExpedientLighting,
-      createExpedientLighting,
-      deleteExpedientLighting
-    } = ExpedientlightingServices()
-    const { getStatus } = StatusLightingServices()
-    const $toast = useToast()
+    const swal = inject('$swal')
+    const showModal = ref(false)
+    const { getAreasByDepartament } = AreaServices()
+    const { getDepartaments } = DepartamentServices()
+    const { getExpedientLighting, createExpedientLighting, deleteExpedientLighting } = ExpedientlightingServices()
+    // const { getStatus } = StatusLightingServices()
+    // const $toast = useToast()
     const expedientLighting = ref([])
-    const statusPublicLighting = ref([])
+    const departaments = ref([])
+    const areas = ref([])
+    // const statusPublicLighting = ref([])
     const isOpen = ref(false)
     const perPage = ref(5)
     const currentPage = ref(1)
@@ -223,33 +241,84 @@ export default {
     const isloading = ref(true)
     const searchValue = ref('')
     const searchField = ref('tarea')
-    const TypeTaskState = ref(false)
-    const DescriptionSolutionState = ref(false)
+    const PublicPlaceState = ref(false)
+    const LocationState = ref(false)
     const HighDateState = ref(false)
-    const addresdescriptionState = ref(false)
-    const DomicileState = ref(false)
-    const LowDateState = ref(false)
+    const departamentState = ref(false)
+    const areaState = ref(false)
+    const DescriptionSolutionState = ref(false)
     const expedientLightingFields = ref({
       expedienteAlumbradoId: 0,
-      tarea: null,
       descripcionSolucion: null,
       fechaAlta: null,
-      fechaBaja: null,
-      domicilio: null,
-      descripcionDomicilio: null,
+      lugarPublico: null,
+      localidad: null,
+      departamentoId: null,
+      areaId: null,
       archivado: false
     })
 
-    getStatus(data => {
-      statusPublicLighting.value = data
+    // getStatus(data => {
+    //   statusPublicLighting.value = data
+    // })
+
+    getDepartaments(data => {
+      departaments.value = data
+      if (data.length === 0) {
+        swal.fire({
+          title: 'No se encuentran departamentos registrados!',
+          text: 'No se encuentran departamentos registrados en el sistema, registre primero un departamento para continuar.',
+          icon: 'warning'
+        })
+      }
     })
 
-    const validateTypeTask = () => {
-      if (!expedientLightingFields.value.tarea) {
-        TypeTaskState.value = false
+    const getAreas = (departamentoId) => {
+      getAreasByDepartament(departamentoId, data => {
+        areas.value = data
+        if (data.length === 0) {
+          swal.fire({
+            title: 'No se encuentran areas registradas!',
+            text: 'No se encuentran areas registradas en el departamento seleccionado, registre primero una area para continuar.',
+            icon: 'warning'
+          })
+        }
+      })
+    }
+
+    const validatePublicPlace = () => {
+      if (!expedientLightingFields.value.lugarPublico) {
+        PublicPlaceState.value = false
         return 'Este campo es requerido'
       }
-      TypeTaskState.value = true
+      PublicPlaceState.value = true
+      return true
+    }
+
+    const validateLocation = () => {
+      if (!expedientLightingFields.value.localidad) {
+        LocationState.value = false
+        return 'Este campo es requerido'
+      }
+      LocationState.value = true
+      return true
+    }
+
+    const validateDepartament = () => {
+      if (!expedientLightingFields.value.departamentoId) {
+        departamentState.value = false
+        return 'Este campo es requerido'
+      }
+      departamentState.value = true
+      return true
+    }
+
+    const validateArea = () => {
+      if (!expedientLightingFields.value.areaId) {
+        areaState.value = false
+        return 'Este campo es requerido'
+      }
+      areaState.value = true
       return true
     }
 
@@ -262,39 +331,12 @@ export default {
       return true
     }
 
-    const validateDomicile = () => {
-      if (!expedientLightingFields.value.domicilio) {
-        DomicileState.value = false
-        return 'Este campo es requerido'
-      }
-      DomicileState.value = true
-      return true
-    }
-
-    const validateLowDate = () => {
-      if (!expedientLightingFields.value.fechaBaja) {
-        LowDateState.value = false
-        return 'Este campo es requerido'
-      }
-      LowDateState.value = true
-      return true
-    }
-
     const validateHighDate = () => {
       if (!expedientLightingFields.value.fechaAlta) {
         HighDateState.value = false
         return 'Este campo es requerido'
       }
       HighDateState.value = true
-      return true
-    }
-
-    const validateAddresdescription = () => {
-      if (!expedientLightingFields.value.descripcionDomicilio) {
-        addresdescriptionState.value = false
-        return 'Este campo es requerido'
-      }
-      addresdescriptionState.value = true
       return true
     }
 
@@ -305,14 +347,26 @@ export default {
 
     const fields = ref([
       { value: 'expedienteAlumbradoId', text: 'ID', sortable: true },
-      { value: 'tarea', text: 'Tarea' },
-      { value: 'fechaAlta', text: 'Fecha Alta' },
-      { value: 'domicilio', text: 'Domicilio' },
-      { value: 'descripcionDomicilio', text: 'Descripcion Domicilio' },
-      { value: 'descripcionSolucion', text: 'Descripcion Solucion' },
-      { value: 'fechaBaja', text: 'Fecha Baja' },
+      { value: 'lugarPublico', text: 'Tarea' },
+      { value: 'localidad', text: 'Fecha Alta' },
+      { value: 'fechaAlta', text: 'Domicilio' },
+      { value: 'departamentoId', text: 'Descripcion Domicilio' },
+      { value: 'areaId', text: 'Descripcion Solucion' },
       { value: 'actions', text: 'Acciones' }
     ])
+
+    const resetexpedientLightingFields = () => {
+      showModal.value = false
+      expedientLightingFields.value = JSON.parse(
+        JSON.stringify(expedientLightingFieldsBlank)
+      )
+      PublicPlaceState.value = false
+      LocationState.value = false
+      HighDateState.value = false
+      departamentState.value = false
+      areaState.value = false
+      DescriptionSolutionState.value = false
+    }
 
     getExpedientLighting(data => {
       expedientLighting.value = data
@@ -350,27 +404,58 @@ export default {
     const addExpedientLighting = () => {
       createExpedientLighting(expedientLightingFields.value, data => {
         refreshTable()
-        $toast.success('Alumbrado registrado correctamente.', {
-          position: 'top-right',
-          duration: 1500
+        swal.fire({
+          title: '¡Expediente publico agregado correctamente!',
+          text: 'Expediente registrado satisfactoriamente',
+          icon: 'success'
         })
+        // $toast.success('Alumbrado registrado correctamente.', {
+        //   position: 'top-right',
+        //   duration: 1500
+        // })
       })
+      showModal.value = false
+      resetexpedientLightingFields()
       // resetStreetLightingFields()
-      expedientLightingFields.value = JSON.parse(
-        JSON.stringify(expedientLightingFieldsBlank)
-      )
+      // expedientLightingFields.value = JSON.parse(
+      //   JSON.stringify(expedientLightingFieldsBlank)
+      // )
     }
 
     const RemoveExpedientLighting = StreetLightingId => {
       isloading.value = true
-      deleteExpedientLighting(StreetLightingId, data => {
-        refreshTable()
-      })
+      swal
+        .fire({
+          title: '¿Estas seguro',
+          text: 'No podras revertir esto',
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonColor: '#3085d6',
+          cancelButtonColor: '#d33',
+          confirmButtonText: 'Si, Archivar Expediente!',
+          cancelButtonText: 'Cancelar'
+        })
+        .then(result => {
+          if (result.isConfirmed) {
+            deleteExpedientLighting(StreetLightingId, data => {
+              refreshTable()
+            })
+            swal.fire({
+              title: '¡Expediente archivado!',
+              text: 'El expediente ha sido archivado satisfactoriamente.',
+              icon: 'success'
+            })
+          } else {
+            isloading.value = false
+          }
+        })
     }
+
     return {
       expedientLighting,
-      statusPublicLighting,
+      // statusPublicLighting,
       expedientLightingFields,
+      showModal,
       isOpen,
       perPage,
       currentPage,
@@ -381,28 +466,29 @@ export default {
       searchField,
       expedientLightingFieldsBlank,
       fields,
-      TypeTaskState,
       DescriptionSolutionState,
       HighDateState,
-      addresdescriptionState,
-      DomicileState,
-      LowDateState,
+      PublicPlaceState,
+      LocationState,
+      departamentState,
+      areaState,
+      areas,
 
       onFiltered,
+      getAreas,
       addExpedientLighting,
       refreshTable,
       RemoveExpedientLighting,
-      validateTypeTask,
       validateDescriptionSolution,
       validateHighDate,
-      validateAddresdescription,
-      validateLowDate,
-      validateDomicile
+      resetexpedientLightingFields,
+      validatePublicPlace,
+      validateLocation,
+      validateDepartament,
+      validateArea
     }
   }
 }
 </script>
 
-<style>
-
-</style>
+<style></style>
